requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
google-api-python-client==2.149.0
python-dotenv==1.0.1
httpx==0.27.2
alembic==1.13.3
lzstring==1.0.4
